LPS-171357 Extract ConvertProcess as components
